move operator logic to pkg/operator
